// src/app/models/client.ts
export class Client {
  nome: string = '';
  apelido: string = '';
  dataNascimento: string = '';
  capacidadeCivil!: number;
  dataEmancipacao: string = '';
  estadoCivil!: number;
  isPCD!: number;
  necessidade: string = '';
  tipoNacionalidade!: number;
  nacionalidade!: number;
  naturalidade!: number;
  sexo!: number;
  nomeMae: string = '';
  nomePai: string = '';
  cpf: string = '';
  situacaoCpf!: number;
  dataConsultaCpf: string = '';
  documentoIdentidade!: number;
  numeroDocumentoIdentidade: string = '';
  orgaoEmissor!: number;
  ufEmissor!: number;
  dataEmissao: string = '';

  pictureSource: string = '';
}